Metric Dictionary

This data frame contains metrics from all tests.

Usage

MetricDictionary

Format

A data frame with 484 rows and 7 variables:

canonicalTestTypeId

The unique id of the canonical test type (character).

testTypeName

The given / common name of the test type (character).

id

PascalCase formatted name of the metric (character).

label

The metric label found in app and cloud UI (character).

label_unit

The metric name returned from the API (character).

header

header of data frame output (character).

units

Metric unit of measure (character).

description

A verbose description of the metric (character).

Source

Generated for demonstration purposes.