Get Test Trials By Teams

Deprecated: Use get_tests instead, which has been expanded to handle all requests.

Get only tests of the specified team for an account.

Usage

get_tests_team(teamId, from = NULL, to = NULL, sync = FALSE, includeInactive = FALSE)

Arguments

teamId

Supply a team’s ID, a list of team IDs, or a string of a comma separated team id’s to receive tests from the specified teams. A maximum of 10 teams can be fetched at once.

from

Optionally supply a time (Unix time stamp) you want the tests from. If you do not supply this value you will receive every test. This parameter is best suited for bulk exports of historical data.

to

Optionally supply a time (Unix time stamp) you want the tests to. If you do not supply this value you will receive every test from the beginning of time or the optionally supplied from parameter. This parameter is best suited for bulk exports of historical data.

sync

The result set will include updated and newly created tests. This parameter is best suited to keep your database in sync with the Hawkin database. If you do not supply this value you will receive every test.

includeInactive

There was a change to the default API configuration to reflect the majority of users API configuration. Inactive tests or tests where active = false are returned in these configuration. Be default, includeInactive is set to FALSE. To return all tests, including disabled trials, set includeInactive to TRUE.

Value

Response will be a data frame containing the trials within the time range (if specified).

Column Name	Type	Description

id	str	Test trial unique ID

active	logi	The trial is active and not disabled

timestamp	int	UNIX time stamp of trial

segment	chr	Description of the test type and trial number of the session (testType:trialNo)

test_type_id	chr	Id of the test type of the trial

test_type_name	chr	Name of the test type of the trial

test_type_canonicalId	chr	Canonical Id of the test type of the trial

test_type_tag_ids	chr	String of Ids associated with tags used during the test trial

test_type_tag_names	chr	String of names of tags used during the test trial

test_type_tag_desc	chr	String of descriptions of tags used during the test trial

athlete_id	chr	Unique Id of the athlete

athlete_name	chr	Athlete given name

athlete_active	logi	The athlete is active

athlete_teams	list	List containing Ids of each team the athlete is on

athlete_groups	list	List containing Ids of each group the athlete is in

All metrics from each test type are included as the remaining variables. If a trial does not have data for a variable it is returned NA.

See also

Examples

if (FALSE) { # \dontrun{
# This is an example of how the function would be called.

## Call for all tests by Group 1
dfteam1 <- get_tests_team(teamId = "team1")

## Call for all tests from Groups 1 & 2
dfteam_1_2 <- get_tests_team(teamId = paste0("team1","team2"))

## Call for all Group 1 tests since a specific date
df_team1_Since <- get_tests_team("team1", from = 1689958617)

} # }