Get Athletes

Get the athletes for an account. Inactive players will only be included if inactive parameter is set to TRUE.

Usage

get_athletes(inactive = FALSE)

Arguments

inactive

FALSE by default to exclude inactive players in database. Set to TRUE if you want inactive players included in the return.

Value

Response will be a data frame containing the athletes that match this query. Each athlete includes the following variables:

id

chr athlete's unique ID

name

chr athlete's given name (First Last)

active

logi athlete is active (TRUE)

teams

chr team ids separated by ","

groups

chr group ids separated by ","

external

chr external ids as strings of "externalName:externalId" separated by ","

Examples

if (FALSE) {
# This is an example of how the function would be called. If you only wish to call active players,
# you don't need to provide any parameters.

df_athletes <- get_athletes()

# If you want to include all athletes, including inactive athletes, include the optional
# `inactive` parameter.

df_wInactive <- get_athletes(inactive = TRUE)

}