Get Test Trials By Teams

Get only tests of the specified team for an account.

Usage

get_tests_team(teamId, from, to, sync = FALSE, active = TRUE)

Arguments

teamId

Supply a team’s or a string of a comma separated list of group id’s to receive tests from specific groups. Recommended to use method paste0(). A maximum of 10 groups can be fetched at once.

from

Optionally supply a time (Unix timestamp) you want the tests from. If you do not supply this value you will receive every test. This parameter is best suited for bulk exports of historical data.

to

Optionally supply a time (Unix timestamp) you want the tests to. If you do not supply this value you will receive every test from the beginning of time or the optionally supplied from parameter. This parameter is best suited for bulk exports of historical data.

sync

The result set will include updated and newly created tests. This parameter is best suited to keep your database in sync with the Hawkin database. If you do not supply this value you will receive every test.

active

There was a change to the default API configuration to reflect the majority of users API configuration. Inactive tests or tests where active:false are returned in these configuration. Be default, active is set to TRUE. To return all tests, including disabled trials, set active to FALSE.

Value

Response will be a data frame containing the trials from the specified team and within the time range (if specified).

id

str Test trial unique ID

active

logi The trial is active and not disabled

timestamp

int UNIX timestamp of trial

segment

chr Description of the test type and trial number of the session (testType:trialNo)

testType.id

chr Id of the test type of the trial

testType.name

chr Name of the test type of the trial

testType.canonicalId

chr Canonical Id of the test type of the trial

athlete.id

chr Unique Id of the athlete

athlete.name

chr Athlete given name

athlete.active

logi The athlete is active

athlete.teams

List list containing Ids of each team the athlete is on

athlete.groups

List list containing Ids of each group the athlete is in

All metrics from each test type are included as the remaining variables. If a trial does not have data for a variable it is returned NA.

Examples

if (FALSE) {
# This is an example of how the function would be called.

## Call for all tests by Group 1
dfteam1 <- get_tests_team(teamId = "team1")

## Call for all tests from Groups 1 & 2
dfteam_1_2 <- get_tests_team(teamId = paste0("team1","team2"))

## Call for all Group 1 tests since a specific date
df_team1_Since <- get_tests_team("team1", from = 1689958617)

}